# adding endianness, radix, 16-nibble groups, and a few tests and tools