* introduced the concept of endianness.  Before this change, the code assumed that the displayed data is big-endian. 
 However, most file formats and most CPU architectures are little-endian.  Thus, this change introduces the support for it.
    The endianness can be controlled by passing bool via `littleEndian`, which defaults to `false` to support the behavior of the previous versions
* introduced ability to group 8 bytes (16 nibbles).  With prevalence of 64-bit computing, the 64-bit (i.e. 8-byte) data is getting more and more popular.
    The 8-byte grouping is enabled by passing "sixteens" into `config.format`
* introduced ability to display the binary data in bases (radixes) other than hexadecimal: binary, octal, decimal and hexadecimal
    The radix is controlled by passing 2, 8, 10 or 16 into `config.radix`
* introduced ability to control if non-printable characters are displayed or replaced with `'.'`.
    To display extended characters, pass `config.extendedChs: true`. The exact behavior of this flag depends on the output type, html or not:
    In `config.html: true` mode, all the characters can be displayed, even 0-0x20 have visual representation.
    In `config.html: false` mode, only the extended characters beyond the end of standard ASCII are displayed.
* implemented and exported `maxnumberlen()` -- calculates how many characters can a number occupy given bittness and radix
* several tweaks improved performance by ~15-30%, depending on the platform, compared to v.0.3.2.  See "perf" chapter below
* a bit more order in the node.js tests:
  * the tests are read from an uniform table.  This allows enumerating the testcases (for quickly locating them when troubleshooting), as well as sharing them with browser tests
  * added ability to do performance tests -- just run `time node test perf`
* enabled browser tests:
  * visual summary with details of all the tests, collapsable and color-coded
  * same set of testcases as in node.js
  * all tests pass now.  Found and fixed a bug that was present in all browsers where they handle bigger-than-byte data differently compared to node.js
* created a static html page to hex display files (view.html)
* restricted the set of node.js versions and browsers (now requires support of `BigInt`: Node.JS 10.4+, browsers since 2018-2020)
* the Travis-ci is passing now
* nits:
  * removed some of unused variables
  * increased formatting consistency

# perf
did a little bit of perf work, trying several alternatives.  Here are the perf results on some of the platforms.  All times here are time it takes to run the first 65 testcases, averaged after 100k iterations (thus JIT is amortized over 100k loops), in comparison with v.0.3.2:

**MacBook Air M1 @3.2GHz, Node.js 16.11.1**  
speedup: 23.8%  
(172-173)μs -> (131-132)μs

**MacBook Pro A1502 i5 5287U @2.9GHz, Node.js 14.15.1**  
speedup: 29.6%  
(686-694)μs -> (483-506)μs  

**Windows 10 Ryzen 5800x @3.8GHz, Node.js 14.16.0**  
speedup: 34.4%  
(450-451)μs -> (295-296)μs

**Windows 10 i7 8550 @ 1.8GHz, Node.js 14.16.0**  
speedup: 32.7%  
(1402-1521)μs -> (944-1078)μs

**Linux 5.4.0 i7-3667U @2GHz, Node.js 10.19.0**  
speedup: 11.4%  
(625-668)μs -> (554-555)μs

**Linux 5.10 R-Pi 4, Node.js 10.24.0**  
speedup: 15.9%  
(2639)μs -> (2219-2382)μs

**Android Pixel 3, Node.js 16.11.1**  
difficult to explain (perhaps cache locality effects), but 100% reproduceable result: 76% speedup  
(2535)μs -> (592)μs

**Android Galaxy Tab S7, Node.js 16.11.1**  
speedup: 33.9%  
(528)μs -> (349)μs


below are individual changes and their impact as observed @ M1.  The end-to-end improvements validated on other platforms (above)
* removing `split().join()` during `*_escape()` saved 40% of the `escape()` function time, or ~2.5% of total time
* moving from `push()` followed by `return ''.concat(parts)` to string `+=` saved ~8.5% of total time (35 -> 32 sec)
* moving from `yield` to splitting the line in the same loop saved ~25% of total time (32 -> 23.5 sec)
* moving from `BigInt` to regular numbers (when BigInt is not needed) saves about 44% of total time (23.5 -> 13 sec).  If BigInt wasn't needed at all, then the perf improvement would've been ~50%
* the constructor is now slower, due to processing more parameters
* streamlining of `toString()` compared to 0.3.2 leads to 18% speedup of the function
* `pad()` -> `num2str()` (essentially, using `"0".repeat(...)`) leads to 4.5x perf improvement within this function
* changing regular expression from `[\000-\040\177-\377]` to `[^\x20-\x7f]` speeds up the replacement by a factor of 27%