@a2800276 2021-11-01
Thanks very much, that was a lot of usefull stuff. Sorry for taking so long!